Documentation, some refactoring and tweaks
